h1st/requirements.txt
# DATA MANAGEMENT & PROCESSING
NumPy >= 1.19.5, < 1.20
Pandas >= 1.2.4
PyArrow >= 4.0.1
Ray[tune] >= 1.4.0; (python_version < '3.9') and (platform_machine != 'armv7l')


# GRAPH PROCESSING
GraphViz >= 0.16


# MACHINE LEARNING
Auto-SKLearn >= 0.12.6; python_version < '3.9'
SciKit-Learn >= 0.24.2
LIME >= 0.2.0.1
SHAP >= 0.39.0

# HYPER-PARAMETER TUNING
Bayesian-Optimization >= 1.2.0
ConfigSpace >= 0.4.19; python_version < '3.9'
Hpbandster >= 0.7.4; python_version < '3.9'
Tabulate >= 0.8.9


# FUZZY LOGIC
SciKit-Fuzzy >= 0.4.2


# MODEL REPOSITORY
CloudPickle >= 1.6.0
PyYAML
    # version range not specified to avoid conflict with AWSCLI
S3FS >= 0.4, < 0.5
ULID-Py >= 1.1.0



# OTHER
Cython >= 0.29.23
PSUtil >= 5.8.0
